+++ packages/x-components/src/components/sort/__tests__/sort-list.spec.ts
@@ -1,79 +1,66 @@
 import type { DeepPartial } from '@empathyco/x-utils'
-import type { RootXStoreState } from '../../../../store/store.types'
+import type { RootXStoreState } from '../../../store/store.types'
 import { mount } from '@vue/test-utils'
 import { describe, expect, it, vi } from 'vitest'
 import { nextTick } from 'vue'
 import { Store } from 'vuex'
-import { XDummyBus } from '../../../../__tests__/bus.dummy'
-import { getDataTestSelector, installNewXPlugin } from '../../../../__tests__/utils'
-import { getXComponentXModuleName, isXComponent } from '../../../../components/x-component.utils'
-import { XPlugin } from '../../../../plugins/x-plugin'
-import { searchXModule } from '../../x-module'
+import { XDummyBus } from '../../../__tests__/bus.dummy'
+import { getDataTestSelector, installNewXPlugin } from '../../../__tests__/utils'
+import { XPlugin } from '../../../plugins/x-plugin'
+import { searchXModule } from '../../../x-modules/search/x-module'
 import SortList from '../sort-list.vue'
-import { resetXSearchStateWith } from './utils'
 
 const bus = new XDummyBus()
 
 function renderSortList({
   template = `
-   <SortList :items="items">
+   <SortList :items="items" :selectedSort="selectedSort">
       <template #default="{ item }">
         {{ item }}
       </template>
-   </SortList>`,
+    </SortList>`,
   items = ['default', 'Price low to high', 'Price high to low'],
   selectedSort = items[0],
 }: Partial<{ template?: string; items?: any[]; selectedSort?: any }> = {}) {
   const store = new Store<DeepPartial<RootXStoreState>>({})
 
-  const wrapper = mount(
+  const parentWrapper = mount(
     {
       template,
       components: { SortList },
-      props: ['items'],
+      props: ['items', 'selectedSort'],
     },
     {
       global: {
         plugins: [installNewXPlugin({ store, initialXModules: [searchXModule] }, bus)],
       },
       store,
-      props: { items },
+      props: { items, selectedSort },
     },
   )
-
-  resetXSearchStateWith(store, { sort: selectedSort })
 
   const onSelectedSortProvided = vi.fn()
   XPlugin.bus.on('SelectedSortProvided', true).subscribe(onSelectedSortProvided)
   const onUserClickedASort = vi.fn()
   XPlugin.bus.on('UserClickedASort', true).subscribe(onUserClickedASort)
 
-  const sortList = wrapper.findComponent(SortList)
+  const sortList = parentWrapper.findComponent(SortList)
 
   return {
     wrapper: sortList,
     onUserClickedASort,
     onSelectedSortProvided,
-    getButton: (index: number) => wrapper.vm.$el.children[index].children[0] as HTMLElement,
+    getButton: (index: number) => parentWrapper.vm.$el.children[index].children[0] as HTMLElement,
     getSelectedItem: () => sortList.get('.x-sort-list__item--is-selected'),
     clickNthItem: async (index: number) => {
       await sortList.findAll(getDataTestSelector('x-sort-button')).at(index)?.trigger('click')
+      await parentWrapper.setProps({ selectedSort: items[index] })
       await nextTick()
     },
   }
 }
 
 describe('testing SortList component', () => {
-  it('is an XComponent', () => {
-    const { wrapper } = renderSortList()
-    expect(isXComponent(wrapper.vm)).toBeTruthy()
-  })
-
-  it('is an XComponent that belongs to the search module', () => {
-    const { wrapper } = renderSortList()
-    expect(getXComponentXModuleName(wrapper.vm)).toBe('search')
-  })
-
   it('allows selecting one of the options of the list', async () => {
     const { getButton, clickNthItem, getSelectedItem, onUserClickedASort } = renderSortList({
       items: ['price', 'relevance', 'offer'],
@@ -85,7 +72,7 @@ describe('testing SortList component', () => {
     expect(onUserClickedASort).toHaveBeenCalledWith({
       eventPayload: 'offer',
       metadata: {
-        moduleName: 'search',
+        moduleName: null,
         target: getButton(2),
         location: 'none',
         replaceable: true,
@@ -102,7 +89,7 @@ describe('testing SortList component', () => {
     expect(onSelectedSortProvided).toHaveBeenCalledWith({
       eventPayload: 'default',
       // This event gets emitted immediately, before the component has been mounted
-      metadata: { moduleName: 'search', location: 'none', replaceable: true },
+      metadata: { moduleName: null, location: 'none', replaceable: true },
     })
   })
 
@@ -110,7 +97,7 @@ describe('testing SortList component', () => {
     it('allows to customize each item using the default slot', () => {
       const { getSelectedItem } = renderSortList({
         template: `
-          <SortList :items="items">
+          <SortList :items="items" :selectedSort="selectedSort">
             <template #default="{ item, isSelected }">
               <span>{{ isSelected }}</span>
             </template>