+++ packages/x-components/src/components/sort/sort-list.vue
@@ -22,13 +22,11 @@
 <script lang="ts">
 import type { Sort } from '@empathyco/x-types'
 import type { Component, PropType } from 'vue'
-import type { VueCSSClasses } from '../../../utils/types'
-import type { XEventsTypes } from '../../../wiring/events.types'
+import type { VueCSSClasses } from '../../utils/index'
+import type { XEvent, XEventsTypes } from '../../wiring/index'
 import { computed, defineComponent, watch } from 'vue'
-import BaseEventButton from '../../../components/base-event-button.vue'
-import { use$x } from '../../../composables/use-$x'
-import { useState } from '../../../composables/use-state'
-import { searchXModule } from '../x-module'
+import { use$x } from '../../composables/use-$x'
+import BaseEventButton from '../base-event-button.vue'
 
 /**
  * Sort list item options.
@@ -45,28 +43,41 @@ interface SortListItem {
  */
 export default defineComponent({
   name: 'SortList',
-  xModule: searchXModule.name,
   components: { BaseEventButton },
   props: {
-    /** The list of possible sort values. */
-    items: {
-      type: Array as PropType<Sort[]>,
-      required: true,
-    },
     /** The transition to use for rendering the list. */
     animation: {
       type: [String, Object] as PropType<string | Component>,
       default: () => 'ul',
     },
+    /** The list of possible sort values. */
+    items: {
+      type: Array as PropType<Sort[]>,
+      required: true,
+    },
+    selectedSort: {
+      type: String as PropType<Sort>,
+      required: true,
+    },
+    selectedSortProvidedEvent: {
+      type: String as PropType<XEvent>,
+      default: 'SelectedSortProvided',
+    },
+    clickedSortEvent: {
+      type: String as PropType<XEvent>,
+      default: 'UserClickedASort',
+    },
   },
   setup(props) {
     const $x = use$x()
 
-    const { sort: selectedSort } = useState('search')
-
-    watch(selectedSort, (value: Sort) => $x.emit('SelectedSortProvided', value), {
-      immediate: true,
-    })
+    watch(
+      () => props.selectedSort,
+      value => $x.emit(props.selectedSortProvidedEvent, value),
+      {
+        immediate: true,
+      },
+    )
 
     /**
      * Sort list items.
@@ -77,16 +88,15 @@ export default defineComponent({
       props.items.map(item => ({
         item,
         cssClasses: {
-          'x-sort-list__item--is-selected': item === selectedSort.value,
-          'x-option-list__item--is-selected': item === selectedSort.value,
+          'x-sort-list__item--is-selected': item === props.selectedSort,
+          'x-option-list__item--is-selected': item === props.selectedSort,
         },
-        event: { UserClickedASort: item },
+        event: { [props.clickedSortEvent]: item },
       })),
     )
 
     return {
       listItems,
-      selectedSort,
     }
   },
 })