+++ packages/x-components/src/x-modules/next-queries/wiring.ts
@@ -162,6 +162,10 @@ export const nextQueriesWiring = createWiring({
     resetNextQueriesWire,
     resetNextQueriesQuery,
   },
+  UserBrowsedToCategory: {
+    resetNextQueriesWire,
+    resetNextQueriesQuery,
+  },
   SelectedRelatedTagsChanged: {
     setNextQueriesRelatedTags,
   },