+++ packages/x-components/src/x-modules/search-box/wiring.ts
@@ -102,6 +102,9 @@ export const searchBoxWiring = createWiring({
   UserClearedQuery: {
     transitionState: setInputStatus('UserClearedQuery'),
   },
+  UserBrowsedToCategory: {
+    transitionState: setInputStatus('UserClearedQuery'),
+  },
   UserFocusedSearchBox: {
     transitionState: setInputStatus('UserFocusedSearchBox'),
   },