+++ packages/x-components/CHANGELOG.md
@@ -3,6 +3,53 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.4.0](https://github.com/empathyco/x/compare/@empathyco/x-components@8.3.0...@empathyco/x-components@8.4.0) (2026-07-16)
+
+### Features
+
+* reset query field when UserBrowsedToCategory (#2139)
+
+
+## [8.3.0](https://github.com/empathyco/x/compare/@empathyco/x-components@8.2.0...@empathyco/x-components@8.3.0) (2026-07-15)
+
+### Features
+
+* **vendor:** result events (#2135)
+
+
+## [8.2.0](https://github.com/empathyco/x/compare/@empathyco/x-components@8.1.1...@empathyco/x-components@8.2.0) (2026-07-14)
+
+### Features
+
+* reset browse fields when UserAcceptedAQuery (#2138)
+
+
+## [8.1.1](https://github.com/empathyco/x/compare/@empathyco/x-components@8.1.0...@empathyco/x-components@8.1.1) (2026-07-14)
+
+**Note:** Version bump only for package @empathyco/x-components
+
+
+
+
+
+## [8.1.0](https://github.com/empathyco/x/compare/@empathyco/x-components@8.0.0...@empathyco/x-components@8.1.0) (2026-07-14)
+
+### Features
+
+* trigger ai suggestions only in no results scenario (#2137)
+
+
+## [8.0.0](https://github.com/empathyco/x/compare/@empathyco/x-components@7.5.0...@empathyco/x-components@8.0.0) (2026-07-13)
+
+### ⚠ BREAKING CHANGES
+
+* Sorting components moved to x-components/src/components/sort. Now, sorting components can be used by the search and browse module by managing them through props.
+
+### Features
+
+* create sort components that could be shared between search and … (#2134)
+
+
 ## [7.5.0](https://github.com/empathyco/x/compare/@empathyco/x-components@7.4.1...@empathyco/x-components@7.5.0) (2026-07-13)
 
 ### Features